[svn-r13580] Description:     Add version # and flags to external link format (as fields in a single byte), in order to accomodate future changes/expansions.
Tested on:
    Mac OS X/32 10.4.9 (amazon)
    Linux/32 2.6 (chicago)
    Linux/64 2.6 (chicago2)
    FreeBSD/32 6.2 (duty)
    FreeBSD/64 6.2 (liberty)

+++ hl/examples/ptExampleFL.c
@@ -72,7 +72,7 @@ int main(void)
  if(err < 0)
      goto out;
 
- printf("Number of packets in packet table after five appends: %d\n", count);
+ printf("Number of packets in packet table after five appends: %d\n", (int)count);
 
     /* Initialize packet table's "current record" */
  err = H5PTcreate_index(ptable);
